\name{DDPM-package}
\alias{DDPM-package}
\docType{package}
\title{
Data Sets for Discrete Probability Models
}

\description{
A wide range of univariate discrete data sets from various applied domains related to distribution theory. The functions allow quick, easy, and efficient access to 100 univariate discrete data sets. The data are related to different applied domains as follows: medical, reliability analysis, engineering, manufacturing, occupational safety, geological sciences, terrorism, psychology, agriculture, environmental sciences, road traffic accidents, demography, actuarial science, law, and justice. The documentation, along with associated references for further details and uses, is presented.
}

\details{
\tabular{ll}{
Package: \tab DDPM\cr
Type: \tab Package\cr
Version: \tab 0.1.0 \cr
Date: \tab 2023-06-14\cr
License: \tab GPL-2\cr
}
}

\author{
 Christophe Chesneau <christophe.chesneau@unicaen.fr>, Muhammad Imran <imranshakoor84@yahoo.com>, M.H Tahir <mht@iub.edu.pk> and Farrukh Jamal <farrukh.jamal@iub.edu.pk>.
}

\section{Maintainers}{
Muhammad Imran <imranshakoor84@yahoo.com>
}





\keyword{Distribution theory}
\keyword{Discrete probability distributions}
\keyword{Count data}
\keyword{Overdispersion}
\keyword{Zero vertex}
\keyword{Zero inflated models}
\keyword{Zero and one inflated models}



